jwt외 세션 두개를 구현중

+++ client/src/hoc/Auth.js
@@ -1,7 +1,7 @@
 import React, { useEffect } from 'react'
 import { useDispatch } from 'react-redux';
 import { useHistory } from 'react-router-dom';
-import { authUser } from '../actions/UserAction';
+import { authUser, oauthUser } from '../actions/UserAction';
 function Auth(Component, option, adminRoute = null) {
     //component => hoc를 적용할 컴포넌트
     //option => null: 아무나 출입,
@@ -12,29 +12,44 @@ function Auth(Component, option, adminRoute = null) {
         const dispatch = useDispatch();
         const history = useHistory();
         useEffect(() => {
-            dispatch(authUser())
-            .then(res => {
-                //로그인하지 않은 상태
-                if(!res.data.auth){
-                    if(res.data.expire){
-                        localStorage.clear();
+            if(localStorage.getItem('access')){
+                dispatch(authUser(localStorage.getItem('access')))
+                .then(res => {
+                    //로그인하지 않은 상태
+                    if(!res.data.auth){
+                        if(res.data.expire){
+                            localStorage.clear();
+                        }
+                        if(option){
+                            return history.push('/login')
+                        }
+                    //로그인 상태
+                    }else{
+                        localStorage.setItem('access', res.data.token);
+                        //로그인 한 상태 중 관리자가 아닌사람이
+                        //adminpage를 들어가려할때
+                        if(adminRoute && res.data.user.role !== 3){
+                            return history.push('/');
+                        }
+                        if(option === false){
+                            return history.push('/');
+                        }
                     }
-                    if(option){
-                        return history.push('/login')
+                })
+            }else{
+                dispatch(oauthUser())
+                .then(res => {
+                    if(!res.data.auth){
+                        if(option){
+                            return history.push('/login');
+                        }
+                    }else{
+                        if(option === false){
+                            return history.push('/');
+                        }
                     }
-                //로그인 상태
-                }else{
-                    localStorage.setItem('access', res.data.token);
-                    //로그인 한 상태 중 관리자가 아닌사람이
-                    //adminpage를 들어가려할때
-                    if(adminRoute && res.data.user.role !== 3){
-                       return history.push('/');
-                    }
-                    if(option === false){
-                        return history.push('/');
-                    }
-                }
-            })
+                })
+            }
         }, [dispatch, history])
 
         return (
